Change: Move last month records login into a component
