+++ src/components/UI/Records/features/LastMonthRecords/index.ts
@@ -1,0 +1,1 @@
+export { LastMonthRecords } from './LastMonthRecords';
